Several improvements to customize the output conversion of attachments.
