+++ Aztec/Classes/Plugin/Plugin.swift
@@ -1,4 +1,5 @@
 import Foundation
+import UIKit
 
 /// Plugin base class.  You can implement a subclass to customize some of the behavior in Aztec.
 ///
@@ -40,6 +41,11 @@ open class Plugin {
     /// When a conversion is not implemented, just return nil.
     ///
     open func convert(_ paragraphProperty: ParagraphProperty) -> ElementNode? { return nil }
+    
+    /// Converts an attachment into the `[Node]`s that represent it.
+    /// When a conversion is not implemented, just return nil.
+    ///
+    open func convert(_ attachment: NSTextAttachment, attributes: [NSAttributedStringKey: Any]) -> [Node]? { return nil }
 
     // MARK: - Equatable
